Move WiFi example files to subdirectory
Renamed and relocated buscar_ip_sumobot.py, wifi.md, and wifiApp.py into códigos_de_ejemplo/WiFi/ for better project organization. No functional code changes were made.
